Add VisionAssistant and refactor GetVision method in OpenAiWrapper
A new property, VisionAssistantId, has been introduced to the Client class in the OpenAiWrapper. The GetVision method has been significantly refactored to utilise this new VisionAssistant for creating and messaging threads, and for creating runs. The method is now asynchronous and returns an OpenAiResponse object that includes potential answers from the OpenAI API.